+++ NEWS.md
@@ -1,5 +1,6 @@
 # runner 0.2.3
 * allow negative lag
+* na_pad in all functions with lag or k
 
 # runner 0.2.2